Fix the way we detect whether window created by user
When windowCreateHandler is triggered, there is no way that we
can decide whether the window was created by user, or it was
created by sync handler because a new window was created from
another device. For the latter case, we should not sync newly
created window to other devices - it isn't a original window, but
a copied window. So in order to decide that, I created a list
called 'windowsAddedBySync'. Before creating a new window in the
sync handler, we push the 'wid' of synced window to the list. In
the callback function of chrome.windows.create, we tag the
previously pushed entry with the window ID. The windowCreateHandler
checks if windowsAddedBySync list has element, and wait until all
elements have been tagged with a window ID. If there is a matching
window ID for the newly created window that triggered
windowCreateHandler, we can decied that the window was created by
sync handler, so we don't need to sync that window with other
devices.