Handle empty url when syncing window/tab
When empty url supplied, the pendingUrl was set to extension's
root url (i.e., chrome-extension://<extension-id>). So if a synced
tab's url is emtpy, set it to "chrome://newtab/". Also, supply
url when syncing a remote created window. Because a new window
has always one tab associated with it, let's just create a tab when
syncing a new window.